requirements.txt
# Core dependencies for Telegram scraping
telethon==1.36.0
python-dotenv==1.0.1

# Data processing and analysis
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.2
jsonlines==4.0.0

# Machine Learning and NLP
torch==2.3.1
transformers==4.41.2
datasets==2.20.0
tokenizers==0.19.1
scikit-learn==1.5.0
seqeval==1.2.2

# Model interpretability
shap==0.45.1
lime==0.2.0.1

# Visualization and reporting
matplotlib==3.9.0
seaborn==0.13.2
plotly==5.22.0

# Jupyter notebooks
jupyter==1.0.0
ipykernel==6.29.4

# Progress tracking
tqdm==4.66.4

# Text processing
regex==2024.5.15

# Evaluation metrics
evaluate==0.4.2